Update run_blip.py
Fixed: the legend/line overflow problem in the diagnostic plots. The issue was that all time segments were being plotted instead of taking a mean over time.
Extant issues: The grid and axes for diag_psd are bugged and won't show up. I'm stumped, no individual component of the plot seems to be responsible, but the overall plot itself isn't much different from the others that currently work.