Point source fix
Tied injection to sph. harm. infrastructure. I'm somewhat concerned about this, and we may want to re-structure it at a later point to use higher lmax than the injection lmax, but in theory it should be fine. (Moreover, it works; a direct pixel-based injection proved much harder than would be expected. There is a way to do the latter, but it would require much more time investment than is currently feasible.)