Added ability to save/load injection/data
Adds functionality to do injection-only runs that save the Injection object and generated h(t) time series data.
Also adds functionality to load the above and run and analysis with all the usual bells and whistles (the additional info carried in the Injection object that is used for post-processing, plotting, etc.)
This will also be useful for running on clusters at scale, so that we can submit high-memory runs to generate our data sets (and only need to do so once) and less-expensive runs for sampling. Useful for benchmarking, makes life easier, and will enable the eventual goal of lowering the numpyro implementation onto GPUs.