Adding injection responses at data frequencies + plotting fixes
Reworked plotting fixes, having resigned myself to "perfect but expensive" as the only workable solution. We now compute and store the responses for all injected signals at the data frequencies, this is useful for plotting and a host of other applications.

This push also includes a prototype dynamic plot sizing that is implemented for diag_spectra() but not fitmaker() (yet)